refactor(4713): decompose voice-runtime giant modules
Split durable announcement metadata and WAV recording lifecycle clusters into
crate-private submodules while preserving the voice module facades.
