+++ ARCHITECTURE.md
@@ -1281,6 +1281,10 @@ src/
 │   ├── secret_file.rs
 │   └── wip_detect.rs
 ├── voice/
+│   ├── announce_meta/
+│   │   └── durable.rs
+│   ├── receiver/
+│   │   └── recording.rs
 │   ├── tts/
 │   │   ├── chunks.rs
 │   │   ├── edge.rs
